requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
openpyxl==3.1.2
Pillow==10.1.0
python-dotenv==1.0.0
gunicorn==21.2.0
